pyproject.toml
[tool.poetry]
name = "nantoken"
version = "0.1.0"
description = "Intelligent LLM Token Tracker - track usage, estimate costs, plan tasks"
authors = ["NanToken Team"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/nantoken"
repository = "https://github.com/yourusername/nantoken"
keywords = ["llm", "ai", "tokens", "budget", "cost", "tracking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.9"
tiktoken = "^0.5.0"
openai = "^1.12.0"
google-generativeai = "^0.3.0"
pyyaml = "^6.0"
tabulate = "^0.9.0"
colorama = "^0.4.6"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
nantoken = "nantoken.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"